+++ inst/rticles/server.R
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/rticles/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2020-12-20
+#> date .: 2021-03-19
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -14,15 +14,6 @@
 #> devtools::install_github("flavjack/inti")
 
 if (file.exists("setup.R")) { source("setup.R") }
-
-library(inti)
-library(shiny)
-library(metathis)
-library(bslib)
-library(shinydashboard)
-library(miniUI)
-library(shinyFiles)
-library(fs)
 
 # -------------------------------------------------------------------------
 # update template ---------------------------------------------------------
@@ -35,7 +26,6 @@ if (FALSE) {
   inti::create_rticles(path = "inst/rticles/template", type = "book")
   file.rename(from = "inst/rticles/template/rticles.Rproj"
               , to = "inst/rticles/template/rticles.proj")
-  
   }
 
 # -------------------------------------------------------------------------