+++ inst/yupana/setup.R
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2021-03-12
+#> date .: 2021-03-19
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -27,19 +27,23 @@ cran <- c(
 
 git <- c("Flavjack/inti")
 
-for (pkg in cran) { 
-  if( !require(pkg, character.only = TRUE) ) {
-    install.packages(pkg)
-    library(pkg, character.only = TRUE)
-  } 
-}
-
-for (pkg in git) { 
-  if( !require(sub(".*/", "", pkg), character.only = TRUE) ) {
-    devtools::install_github(pkg, upgrade = TRUE)
-    library(sub(".*/", "", pkg), character.only = TRUE)
-  } 
-}
+suppressPackageStartupMessages({
+  
+  for (pkg in cran) { 
+    if( !require(pkg, character.only = T) ) {
+      install.packages(pkg)
+      library(pkg, character.only = T)
+    } 
+  }
+  
+  for (pkg in git) { 
+    if( !require(sub(".*/", "", pkg), character.only = T) ) {
+      devtools::install_github(pkg, upgrade = T)
+      library(sub(".*/", "", pkg), character.only = T)
+    } 
+  }
+  
+})
 
 rm(cran, git, pkg)